requirements.txt
PyMuPDF
sentence-transformers
faiss-cpu
transformers
torch
numpy
gradio
pytest
tqdm
